Close #1002 - xsl transformation of the XML when using SMTPSend step
Javascriptable xslFilepath and contentType properties has been added.

If no source selected, the current outputDocument is used, even if no xsl is defined (act as pretty print).

The xsl is apply to the source results (concatened if more than one) or the outputDocument if no source.

The contentType property can be used to force the content type of the mail.

 * without xsl : text/plain
 * with xsl : text/html
 
 The documentation must be reviewed in #3036

Former-commit-id: b1ac8d90860b3ab0751ac51472b6a36e77d4200a
Former-commit-id: 6ccc9c0046770a37abdcb4fa2f70195f40b9981a
Former-commit-id: 9ced60e91d0068b4650f06dbcb677d86a8e70cdc