Close #3188 - Order parallel step response
Element are now ordered.

Execute from attached project the sequence !WhatTimeParallel and the order must be 8/2/4.

Execute the sequence !IteratorParallel with the test case X3 and X10, the order must be the same as the first part of the XML.

Try in an older CEMS version to compare.

A case doesn't work with the Iterator + Parallel case. Only the first output true step is ordered (set ouput false on the Serial of IteratorParallel). We cannot solve this issue with the current sequencer structure.

Former-commit-id: 98a864e20a74bd00ea5c074110797981e466e070
Former-commit-id: 79fdd97d1a5cc24d41c2b96a3b8a9318df748619
Former-commit-id: c8e64d4327b24f50eaba9c8eaf2bb4b072fc6395