Close #3188 - Order parallel step response
Element are now ordered.

Execute from attached project the sequence !WhatTimeParallel and the order must be 8/2/4.

Execute the sequence !IteratorParallel with the test case X3 and X10, the order must be the same as the first part of the XML.

Try in an older CEMS version to compare.

A case doesn't work with the Iterator + Parallel case. Only the first output true step is ordered (set ouput false on the Serial of IteratorParallel). We cannot solve this issue with the current sequencer structure.

Former-commit-id: 639777a0f16212c2e145f94a15323cd70b89db44
Former-commit-id: dfb56d678ff09a5c810c9838424705634ea057c3
Former-commit-id: d9b5379b0473d9d9b08c2b062b0f9fb5acd08231