Close #3299 - Add XVNC port Number in Administration summary page .
Done.

The port can be see in the "Java system properties" button on the mail admin page.

This is very specific and doesn't require a dedicated conditional (Linux +DISPLAY env) IHM in the admin page.

A version will be deployed on http://c8o_dev.convertigo.net as soon as the build is available.

Former-commit-id: 457641fed9e76cb76da9ca5ec3587f53793f4217
Former-commit-id: 819c0b4970d5a5aa7ca3cbc28d832a2fd04e726d
Former-commit-id: 76261cf7d63008551c323c2ce98081444e0ad93b