Close #3299 - Add XVNC port Number in Administration summary page .
Done.

The port can be see in the "Java system properties" button on the mail admin page.

This is very specific and doesn't require a dedicated conditional (Linux +DISPLAY env) IHM in the admin page.

A version will be deployed on http://c8o_dev.convertigo.net as soon as the build is available.

Former-commit-id: 986cbf4dd3e11d1abec565bb9f8f19691b123f62
Former-commit-id: 568cc24d50859295987098014e24c0c5c9a915df
Former-commit-id: a74337e7850b03ba71787e21cb91bace2fe113b5