Close #3082 - Strange behaviors arount SMTP send and Copy steps
Handle the case of empty nodeList to copy with the !XmlCopyStep that cause the step_id attribute.

Handle the case of copy of copy that cause the step_copy attribute.

The Ticket3082 ticket has 2 sequences. One with a bad xpath (cause an empty nodelist) for !XmlCopyStep and another with a valid xpath (generate result).

The darty project should be tested to confirm that.

Done in trunk and 6.2.x branch.

Former-commit-id: 4e013e266a44ff2cf9a6986d7e074a7ea55d6344
Former-commit-id: adb53823b8d7606ebc74346ba54044f778fb8db6
Former-commit-id: 2b286a5ec321fd99652bd815ceb7e374595792d5