Close #3082 - Strange behaviors arount SMTP send and Copy steps
Handle the case of empty nodeList to copy with the !XmlCopyStep that cause the step_id attribute.

Handle the case of copy of copy that cause the step_copy attribute.

The Ticket3082 ticket has 2 sequences. One with a bad xpath (cause an empty nodelist) for !XmlCopyStep and another with a valid xpath (generate result).

The darty project should be tested to confirm that.

Done in trunk and 6.2.x branch.

Former-commit-id: 0e383e49daedcd27ed9d8f852a2b1837957c4188
Former-commit-id: e7323b9754f390102d76fa816aa7e4eb77432008
Former-commit-id: 71629e7e320b665fa0d32fe125b58d87ac2f01a1